+++ htimeseries/__init__.py
@@ -1,4 +1,4 @@
 from .htimeseries import *  # NOQA
-from .tzinfofromstring import *  # NOQA
+from .timezone_utils import *  # NOQA
 
 __version__ = "0.1.0.dev0"